Added PKCS#11 multi-lib support for pageant.
This is still work in progress. I did some initial testing, and it seems to be working fine. I still have to do some UI changes to be able to add library dll.

Now, when PKCS#11 library is loaded all available certificates are displayed in key list, and any can be used for authentication.